Add postprocessing step to ensure RoleBindings are created correctly
The builtin helm_namespace postprocessing filter changes the
`.metadata.namespace` field of all objects in the given directory. The
RoleBinding `cert-manager-webhook:webhook-authentication-reader`,
however must exist in namespace `kube-system`, and not whichever
namespace cert-manager is installed into.

This commit provides an extra postprocessing filter which edits this
RoleBinding and makes sure that it is installed in namespace
`kube-system`.