Use external public DNS server
Relying on the clusters DNS resolution can be none functional. E.g when
there is a private and public zone for the clusters base domain. This is
the case with OpenShif4 on GCP (and also on other clouds). cert-manager
was creating the TXT record in the wrong zone. Setting this option makes
cert-manager to detect the correct zone.

On top of this, it can also be beneficial for all other clusters. Using
a external view might help prevent other issues within the DNS system.