Implement acme-dns self-registration
This commit extends the component with support for self-registration on
an existing acme-dns instance. The registration itself is done by script
`component/acme-dns-scripts/register.sh`. Regular checking of the client
credentials (and reregistration if the credentials don't function) is
done by script `component/acme-dns-scripts/check.sh`.

The shell scripts in `component/acme-dns-scripts` are deployed to the
cluster in a configmap. The shell scripts expect their arguments as
environment variables.

The component configures a Job which performs the initial registration
and a CronJob which verifies the credentials every 24h.

The exact cronjob execution time is randomized between midnight and 2am,
to avoid load spikes on the acme-dns API if many clusters are using the
same acme-dns instance.