Fix DNS01 challenge self-check
We specify custom recursive nameservers for cert-manager to use when
validating DNS01 challenges. However, in some cases, cert-manager will
still try to connect to the DNS01 DNS server directly, unless the
command line argument `--dns01-recursive-nameservers-only` is given as
well.

Since the intention of the config is to be able to specify exactly which
nameservers cert-manager should use, we need to also give
`--dns01-recursive-nameservers-only` to ensure that cert-manager always
uses the configured recursive nameservers.