Make acme-dns registration job an ArgoCD sync hook
Updating plain Jobs is basically impossible, so we run the registration
job as an ArgoCD hook and ensure it gets deleted when it succeeds. The
registration script should always be a no-op when acme-dns credentials
already exist for in the secret.

To ensure the old registration job is deleted and we don't get conflicts
when trying to run the new hook job, we change the registration job's
name to `create-acme-dns-client`.