Move namespace management to component
Until now, we've used the `helm_namespace` builtin postprocessing filter
to manage the cert-manager namespace (and add `metadata.namespace` to
manifests). With ArgoCD, there's no need to inject the namespace to the
manifests rendered from the Helm chart.

This commit removes the call to the superfluous `helm_namespace`
postprocessing filter, and moves management of the namespace manifest to
the component proper.

This change also prepares the component to allow us to add custom labels
and annotations to the namespace.