Fix acmeClient not correctly translated to solver
When using an acmeClient and referencing it in an issuer, the solver configuration
for the issuers solver is not correctly set if you want to use wildcard domains.
This fix will allow defining wildcard domains in the acmeClients to correctly set
the solver config, but will strip the wildcard part from the configuration of the
acmeClient configuration.
Unfortuneatly this will make using legacy `acme_dns_api` parameter not backward
compatible anymore.