Switch to server-side apply for acme-dns registration script
This is necessary now that we use server-side apply for the ArgoCD app.

If the registration script uses cliend-side apply, there's a race
condition during the first sync where it's possible that ArgoCD's
server-side apply overwrites the `acmedns.json` written by the
registration sync job.

With server-side apply in the registration script, we can also switch
back to just rendering a partial `Secret` resource instead of injecting
the `acmedns.json` data into the secret we read from the cluster.