Ensure cert-manager pod is restarted on config changes
We add a postprocessing filter which hashes the cert-manager `config`
Helm value and sets an annotation with the hash value on the
cert-manager pod template.

The annotation is only added if the config isn't empty, since the Helm
chart will remove the configmap mount when the config is empty, which
triggers a pod restart anyway.